refactoring: moved endpoint code into separate submodule
