+++ src/bia_bob/_utilities.py
@@ -68,6 +68,10 @@ def generate_response_to_user(model, user_prompt: str, image=None, additional_sy
 
 def generate_response(chat_history, image, model, system_prompt, user_prompt, vision_system_prompt):
     from ._machinery import Context
+    from .endpoints._openai import generate_response_from_openai
+    from .endpoints._googlevertex import generate_response_from_vertex_ai
+    from .endpoints._anthropic import generate_response_from_anthropic
+
     if Context.endpoint is not None:
         full_response = generate_response_from_openai(model, system_prompt, user_prompt, chat_history, image,
                                                       base_url=Context.endpoint, api_key=Context.api_key,
@@ -310,188 +314,6 @@ def is_notebook() -> bool:
         return False      # Probably standard Python interpreter
 
 
-def generate_response_from_anthropic(model: str, system_prompt: str, user_prompt: str, chat_history, image=None,
-                                  base_url:str=None, api_key:str=None, vision_model:str = None, vision_system_prompt:str = None):
-    # e.g. claude-3-5-sonnet-20240620 or claude-3-opus-20240229
-    from anthropic import Anthropic
-    from ._machinery import Context
-
-    if image is None:
-        if Context.client is None or not isinstance(Context.client, Anthropic):
-            Context.client = Anthropic(base_url=base_url, api_key=api_key)
-        system_message = system_prompt
-        user_message = [{
-                    "role": "user",
-                    "content": user_prompt,
-                }]
-        client = Context.client
-    else:
-        if Context.vision_client is None or not isinstance(Context.vision_client, Anthropic):
-            Context.vision_client = Anthropic(base_url=base_url, api_key=api_key)
-        system_message = vision_system_prompt
-        user_message = image_to_message_claude(image, user_prompt)
-        client = Context.vision_client
-        model = vision_model
-
-    response = client.messages.create(
-        messages=chat_history + user_message,
-        system=system_message,
-        model=model,
-        max_tokens=4096,
-    )
-    reply = response.content[0].text
-
-    assistant_message = [{"role": "assistant", "content": reply}]
-
-    Context.chat += user_message + assistant_message
-
-    return reply
-
-
-def generate_response_from_openai(model: str, system_prompt: str, user_prompt: str, chat_history, image=None,
-                                  base_url:str=None, api_key:str=None, vision_model:str = None, vision_system_prompt:str = None):
-    """A prompt helper function that sends a message to openAI
-    and returns only the text response.
-    """
-    from openai import OpenAI
-    from ._machinery import Context
-
-    # assemble prompt
-    user_message = [{"role": "user", "content": user_prompt}]
-    image_message = []
-    kwargs = {}
-    model_init_kwargs = {}
-
-    if api_key is not None:
-        model_init_kwargs['api_key'] = api_key
-    if base_url is not None:
-        model_init_kwargs['base_url'] = base_url
-
-    if image is None: # normal text-based prompt
-        system_message = [{"role": "system", "content": system_prompt}]
-
-        # init client
-        if Context.client is None or not isinstance(Context.client, OpenAI):
-            Context.client = OpenAI(**model_init_kwargs)
-        client = Context.client
-    else:
-        system_message = [{"role": "system", "content": vision_system_prompt}]
-
-        if 'llava' in vision_model:
-            system_message = "" # llava crashes when the system prompt is too long
-            image_message = image_to_message_llava(image, user_prompt)
-            user_message = []
-        else:
-            image_message = image_to_message(image)
-
-        if vision_model == 'gpt-4-vision-preview':
-            # this seems necessary according to the docs:
-            # https://platform.openai.com/docs/guides/vision
-            # if it is not provided, the response will be
-            # cropped to half a sentence
-            kwargs['max_tokens'] = 3000
-
-        if Context.vision_client is None or not isinstance(Context.vision_client, OpenAI):
-            Context.vision_client = OpenAI(**model_init_kwargs)
-        client = Context.vision_client
-        model = vision_model
-
-    if Context.seed is not None:
-        kwargs['seed'] = Context.seed
-    if Context.temperature is not None:
-        kwargs['temperature'] = Context.temperature
-
-    if Context.verbose:
-        print("messages=", system_message + chat_history + image_message + user_message)
-
-    # retrieve answer
-    response = client.chat.completions.create(
-        messages=system_message + chat_history + image_message + user_message,
-        model=model,
-        **kwargs
-    )  # stream=True would be nice
-    reply = response.choices[0].message.content
-
-    # store question and answer in chat history
-    assistant_message = [{"role": "assistant", "content": reply}]
-
-    Context.chat += user_message + assistant_message
-
-    return reply
-
-
-def generate_response_from_vertex_ai(model: str, system_prompt: str, user_prompt: str, chat_history, image=None, vision_model:str = None, vision_system_prompt:str = None):
-    """A prompt helper function that sends a message to Google Vertex AI
-    and returns only the text response.
-
-    See also
-    --------
-    https://colab.research.google.com/github/GoogleCloudPlatform/generative-ai/blob/main/gemini/getting-started/intro_gemini_python.ipynb#scrollTo=SFbGVflTfBBk
-    """
-    # from vertexai.generative_models._generative_models import ChatSession
-    from ._machinery import Context
-    from vertexai.preview.generative_models import (
-        GenerationConfig,
-        GenerativeModel,
-        Image,
-        Part,
-        ChatSession,
-    )
-
-
-    # if "vision" in Context.model:
-        # We need to do some special case here, because the vision model seems to not support chats (yet).
-
-
-    if image is None:
-        if Context.client is None or not isinstance(Context.client, ChatSession):
-            gemini_model = GenerativeModel(model)
-            Context.client = gemini_model.start_chat()
-            system_result = Context.client.send_message(
-                system_prompt + "\n\nConfirm these general instructions by answering 'yes'.").text
-
-        if len(system_prompt) > 0:
-            system_prompt = create_system_prompt(reusable_variables_block="")
-
-        prompt = f"""
-                   {system_prompt}
-                   
-                   # Task
-                   This is the task:
-                   {user_prompt}
-                   
-                   Remember: Your output should be 1) a summary, 2) a plan and 3) the code.
-                   """
-
-        response = Context.client.send_message(prompt).text
-
-    else: #if image is not None:
-        if Context.client is None or not isinstance(Context.client, GenerativeModel):
-            Context.vision_client = GenerativeModel(vision_model)
-
-        from stackview._image_widget import _img_to_rgb
-
-        rgb_image = _img_to_rgb(image)
-        byte_stream = numpy_to_bytestream(rgb_image)
-
-        image = Image.from_bytes(byte_stream)
-
-
-        prompt = f"""
-               {vision_system_prompt}
-
-               # Task
-               This is the task:
-               {user_prompt}
-               """
-
-        prompt = [image, prompt]
-
-        response = Context.vision_client.generate_content(prompt).text
-
-    return response
-
-
 def numpy_to_bytestream(data):
     """Turn a NumPy array into a bytestream"""
     import numpy as np
@@ -510,60 +332,6 @@ def numpy_to_bytestream(data):
     # return the beginning of the file as a bytestream
     bytes_io.seek(0)
     return bytes_io.read()
-
-
-def image_to_message(image):
-    import base64
-
-    from stackview._image_widget import _img_to_rgb
-
-    rgb_image = _img_to_rgb(image)
-    byte_stream = numpy_to_bytestream(rgb_image)
-    base64_image = base64.b64encode(byte_stream).decode('utf-8')
-
-    return [{"role": "user", "content": [{
-        "type": "image_url",
-        #"image_url": f"data:image/jpeg;base64,{base64_image}",
-        # from: https://platform.openai.com/docs/guides/vision
-        "image_url": {
-            "url": f"data:image/jpeg;base64,{base64_image}"
-        }
-
-    }]}]
-
-
-def image_to_message_llava(image, prompt):
-    import base64
-
-    from stackview._image_widget import _img_to_rgb
-
-    rgb_image = _img_to_rgb(image)
-    byte_stream = numpy_to_bytestream(rgb_image)
-    base64_image = base64.b64encode(byte_stream).decode('utf-8')
-
-    return [{
-        'role': 'user',
-        'content': prompt,
-        'images': [base64_image]
-    }]
-
-
-def image_to_message_claude(image, prompt):
-    import base64
-
-    from stackview._image_widget import _img_to_rgb
-
-    rgb_image = _img_to_rgb(image)
-    byte_stream = numpy_to_bytestream(rgb_image)
-    base64_image = base64.b64encode(byte_stream).decode('utf-8')
-
-    return [{
-            "role": 'user',
-            "content": [
-                {"type": "image", "source": {"type": "base64", "media_type": "image/png", "data": base64_image}},
-                {"type": "text", "text": prompt}
-            ]
-        }]
 
 
 def is_image(potential_image):
